backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
databases[postgresql]==0.9.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pillow==10.1.0
PyPDF2==3.0.1
ebooklib==0.18
beautifulsoup4==4.12.2
nltk==3.8.1
google-generativeai==0.3.2
pydantic==1.10.12
httpx==0.25.2
# TTS related packages - these will be installed via conda and tortoise-tts
tqdm
rotary_embedding_torch
inflect
einops==0.4.1
unidecode
scipy
librosa==0.9.1
numpy
numba
progressbar
threadpoolctl
llvmlite
appdirs
psutil
sounddevice 